Add a `qs` bin and `CLI` class for it's logic
This adds a `qs` bin and a `CLI` for it's logic. This is intended
for starting a qs daemon that will process background jobs. It
uses CLIRB and layers on it's own behavior. This is mostly just
a starting point to work from and will be built on as the daemon
process logic is built out.