Implement `Config` for evaluated config files
This implements the `Config` that is used for evaluating the
user provided configuration files. These are similar to rackup
files and will provide all the options needed to run a Qs daemon
process. The intent is for users to require in their queues and
then build a daemon from the one they want to run. Calling `run`
and passing it their built daemon will trigger Qs to start a
process for running jobs from the provided queue.