Add daemon configuration, provides options and validation
This follows the Deas pattern for a server's configuration and
add's a configuration class for a `Daemon`. This provides the
daemon options and will validate the configuration. This will
allow user's to customize a daemon and also ensure that their
configuration and queue will run as expected. This provides a
friendlier experience than it erroring later after the process
has already started and tries to run a job.