Update daemon config and `Process` interaction
This cleans up and formalizes some of the `Daemon` interface,
specifically where it interacts with the `Process`. This was
originally hacked in to make the `Process` work as needed. The
options that the `Process` class uses are now set using the DSL
and others are properly pulled from the `Queue`. This revealed
that the `queue`, `init_procs` and `error_procs` for a daemon's
configuration were not being copied down from the class to the
instance correctly. This sets up updates to the signal handling,
which affects the `Process` class.