Use dat-worker-pool's spy in the daemon tests
This removes the custom dat-worker-pool spy and uses the one
provided by the gem. This also tweaks the testing to not build
an actual `DatWorkerPool` instance. This had the effect of `Qs`
testing `DatWorkerPool`'s behavior, which isn't necessary. This
switches to only testing `Qs` logic and relying on the spy.