Simply CLI `Config` class
This simplifies the CLI `Config` class by removing the extra
`parse` method and making it's validation clearer. Instead of
lazily validating when the `daemon` method is called, this
validates as part of building the class. This is a more
straightforward workflow and doesn't hinge on calling the
`daemon` method. The `parse` method was superfulous and this
reduces the "code debt" of Qs by removing it.