Update and simplify `Process`
This updates and simplifies the `Process` class. It removes all
the signal logic it had as this is now handled by the
`ProcessSignal`. This also removes the `PIDFile` logic that was
an internal helper in the file. It is now its own separate file
that it just requires. The `Process` is now only about running
a daemon and handling restarting it.

This also removes the tmp process file and switches the cli to
using the real one. This was a temporary placeholder while logic
was being reworked.