Change `Daemon` module organization
This changes the `Daemon` module organization and changes how its
modules are defined. The file now opens both `Qs` and the `Daemon`
module which changes the amount of leading spaces for almost every
line in the file. This also reorganizes the modules to match how
`Sanford` organizes its modules for its `Server`. This is to make
them similar and easier to see the similarities and differences
between the libraries. This doesn't change any functionality, just
changes the organization of the code.

/cc @kellyredding