Implement `Route#run`
This adds the logic for running a route. It will build and run a
`QsRunner`, passing the job, the job params and the daemon data's
logger to it. The route was implemented before the runner or job
so it was left as a todo. This is needed to run a job handler when
a job is pulled off of a queue.