Add `enqueue` to `Qs` and `add` to `Queue`
This adds an `enqueue` method to `Qs` and `add` to `Queue`. These
allow adding jobs to queues in redis to be processed by daemons.
The `enqueue` method on `Qs` takes a queue name, a job name and
job params. The `add` method on `Queue` only takes a job name and
job params. This allows multiple convenient methods to add jobs
to a queue to be processed.