Add system tests for `Daemon`
This adds system tests that test running a `Daemon` in a separate
thread and adding jobs to it. This ensures that from the `Daemon`
to the `JobHandler` jobs are processed correctly. This includes
properly handling exceptions that are raised while processing a
job.

This requires `JobHandler` from `Qs` because it is not being
required via any other `Qs` file.