Remove `TmpDaemon`, use `Daemon` instead
This removes the temporary placeholder `TmpDaemon` that was used
in place of `Daemon` and switches to using `Daemon` directly. This
was needed for the `ConfigFile` and `Process` related parts of
Qs. They were implemented using the `TmpDaemon` but now that the
`Daemon` has been implemented it can be used instead. This will
allow qs to actually run a daemon and process jobs.