Properly create/close IO pipes on start/shutdown
This updates the `Daemon` to properly handle its IO pipes for
signalling when a worker is available or a queue is empty. The test
suite was randomly failing because the process opened too many files
(IO pipes) and didn't close them. This fixes the issue by wrapping
the IOs in an internal class, `IOPipe`. `IOPipe` has a `setup`
and `teardown` method which will build the IOs and close them. This
is then called when a daemon is started and stopped. This error only
showed up in the test suite because it built many daemon instances.
In a typical scenario, you would only run a single daemon instance
for a process so this error wouldn't occur.