Add benchmarking script
This adds a script for benchmarking adding and running jobs in
Qs. The script adds a large number of jobs and then runs them,
benchmarking how long it takes to do both actions. This is to
provide insight if changes have significantly slowed down Qs. It
also serves as a simple script and configuration for running a
Qs daemon and making sure everything works as expected.

Because of Qs asynchronous nature its difficult to get an
accurate time for how long it takes to run a set of jobs. This
uses a pipe to communicate progress between process which has
I/O overhead. In short this means that no matter what the job
does, there is a cap on how fast Qs performs in this script. Thus,
this should not be used to measure how fast/slow Qs is and should
only be used as a guide if changes are significantly slowing down
Qs.