Add `TestClient` and test mode, track enqueued jobs in memory
This adds a `TestClient` that is built when Qs is in test mode (set
with an env variable). The test client is a helper for tests that
will track any jobs that are enqueued in memory and allow
inspecting them. The jobs are stored on the queue that they are
enqueued on. This is nice for testing that a job was enqueued
and making sure it was enqueued with the correct params.

This also includes a `reset!` method on the `Qs` module. This
allows undoing its current client and redis connection so `init`
can be called again. This will allow switching between live and
test mode in the same process if desired (like qs own test suite).

This also includes a minor cleanup to the daemon tests that moves
some "setup" logic into the setup block instead of the should
block.