Add `TestHelpers` for mixin for building a runner/handler
This adds a `TestHelpers` mixin for use in a test suite. It
provides convenient methods for building a test runner or a
handler. These can be used to test job handler

This also commonizes a `RunnerSpy` from the route tests so it can
be reused in the test helper tests. Finally, this removes a "spy"
class that was in the `test/support` folder but wasn't being used.
This was something I randomly played with for tracking method
calls before we had `Assert.stub`.