Allow configuring a serializer/deserializer
This updates Qs to allow users to configure a custom serializer
and deserializer. By default, they serializer and deserializer use
`JSON` to convert the jobs to a string and back. This allows users
more flexibility in choosing their dependencies. The only
change in functionality is that running a daemon will now require
`Qs` to be initialized. Other then that, no other functionality has
changed and it allows using custom serialization instead of being
forced into using JSON and Oj.

This adds a `serialize` and `deserialize` method to the `Qs`
module. This is used through Qs to handle the adding and processing
jobs but can also be helpful for debugging. It provides easy access
to the configured serializer and deserializer.

The `JSON` gem is not required by default and must be brought in
for the defaults to work. This is for ruby 1.8.7 compatibility.
For ruby 1.9 and later the `JSON` module just needs to be required
from the standard library.

Finally, this also adds a few manual sleeps to the daemon tests.
In rare cases the daemon threads would not get enough time on the
processor to run as far as the tests expected. This caused tests
to randomly fail. This just adds a small delay and tries to force
ruby to run the thread longer.