Add job timeouts
This adds job timeouts that can be configured globally for all jobs
or on individual jobs. This allows controlling run-away jobs that
will never end. If a timeout value is given, then `SystemTimer`
is used to ensure the job doesn't run past the value. A
`Qs::TimeoutError` is thrown if it does. If a timeout value isn't
given, then the job will be allowed to run indefinitely. Finally,
a specific timeout for a handler will be used over the globally
configured timeout.