Expand `ErrorHandler` to take more information
This updates the `ErrorHandler` to take more information about the
job that failed. This includes passing the queue that the job
came from, the serialized payload of the job and the handler
class that was being used to run it. The handler class can be `nil`
if the error happened before a handler class was found.

These are being added to expand the functionality of the error
handler and to help with handling errors that happen before a job
is deserialized. Specifically, this is setup for changing how the
hard shutdown feature works. A new version of dat worker pool will
raise an exception on worker threads that are running jobs when it
attempts to force a shutdown. This means an exception can happen
at any point, even before a job has been deserialized. For this
case, we still want to get the serialized payload so we can add
the job back to the queue because it never had a chance to run.
By passing the queue key and serialized payload to the error
handler this will be possible.

This also updates to the latest assert to keep up-to-date and to
use its latest features.