Expand redis item to track the state of processing it
This expands the redis item to not be a simple struct and to
instead track all of the state of processing it. This includes the
job and handler class that were used to run it (if we were able to
deserialize it and route it), the exception that it generated
(again, only if one was thrown) and the time it took to run it.
This is setup for a new dat worker pool that better handles hard
shutdowns and allows custom error handling.

By making the redis item track its state, we can add error handling
to dat worker pool. The error handling can be smart and use the
redis items state to make decisions. For example, if a job was
never started, we can requeue it. This ensures Qs can do its best
to not "lose" any jobs because of errors, specifically when its
being forced to shutdown.

This also changes the error context to change a queue redis key to
a queue name. In general, the redis key is intended to be internal,
so the user-defined error procs shouldn't be working with it
directly.