Add `push` and `prepend` to `Client`
This adds a `push` and `prepend` method to clients. This will be
used to requeue redis items. If a redis item fails before it has
been started, or if a redis item is left on the dat worker pool
when its shutdown, we want to requeue this work before exiting. To
facilitate this, clients now have low-level push and prepend
methods that will take the raw components of a redis item: a queue
redis key and a serialized payload. This will either add the item
to the front or back of the queue.

The `push` and `prepend` methods were not added to the `Qs` module
because they aren't intended for "public" use. The components they
require are not easily available to end-users and require knowledge
of how Qs works to build properly.