Update `Daemon` to build a client instead of a redis connection
This updates the `Daemon` to build a client instead of a redis
connection directly. This is setup for adding a new dat worker
pool. The new dat worker pool changes the shutdown process and
Qs will requeue jobs that were forced to shutdown and didn't
get started and it will also requeue any work left on dat worker
pool's queue. To facilitate this, its better for the daemon to
have a client so it can add jobs to a queue the same way it is
done elsewhere. This also builds the client with enough
connections for each worker to need to requeue a job.

This also fixes an issue with "queue starvation" by shuffling the
keys that are passed when pulling work off a queue. If this isn't
done, then the same queue is always first in the list which means
any queues behind it may never run jobs if it is always full. By
shuffling them they all get a chance to have work pulled from
them.