Update to the latest `DatWorkerPool`
This updates to the latest `DatWorkerPool` which includes on worker
error callbacks and better shutdown behavior. The worker pool is
now configured with an on error callback and is always shutdown,
even when halting. This is part of an effort to better handle
shutting down with a timeout and ensuring that jobs aren't lost.

An on-error callback is configured on the worker pool to catch any
errors that happen outside running the payload handler. The only
known use-case for this is when the worker pools shutdown times
out. In this case, the worker pool will raise errors on its workers
threads. These errors can be raised at anytime, which means they
can occur outside running a payload handler. The on-error callback
will requeue any jobs that the worker pulled off but didn't start
processing. If the worker started processing the job, then the
on-error callback won't receive the exception because the payload
handler will catch it and do the normal error handling logic.

The daemon shutdown behavior is also being changed with the new
dat worker pool. The worker pool is now always shutdown. The only
difference is if the daemon is halted, it will shutdown the worker
pool immediately instead of using the configured timeout. Once the
worker pool has shutdown, the daemon will also requeue any redis
items left on the pool. This is being done instead of the previous
logic that tried to wait until the queue emptied.

Finally, this also fixes a minor issue where the `NULL` IO pipe
could get closed. This happened in the test suite and caused
random tests to fail.