Hotfix: Job's payload, convert name to a string
This is a small change that manually calls `to_s` on a job's name
when its converted to a payload. This is similar to how the job's
params are stringified when building its payload. This, and
stringifying the params, is to avoid annoying serialization errors
for symbols.

This also includes minor style cleanups lining up a jobs name,
params and created at.

/cc @kellyredding