Test helpers, serialize when enqueuing and building test runner
This updates the testing tools to serialize when enqueuing jobs
with the test client and when building a test runner. This ensures
the jobs being enqueued are valid and will work when enqueuing
them in production. This also ensures the params being used to
test a handler are valid and are what can be expected in
production. This is to improve the testing tools.