Pull out `IOPipe`, setup for use in the `Process`
This pulls out the `IOPipe` so that it can be used in the
`Process`. This slightly changes its API as well and updates
the `Daemon` to work with the changes. All of this is setup for
changing how Qs handles unix signals.

In a future PR, the `Process` will instead of joining the daemons
thread to the main thread, it will start a loop where it listens
for signals on an `IOPipe`. The signal traps will be changed to be
super simple and only write to the `IOPipe`. The benefit of this is
its a best practice to limit what is done in a signal trap as much
as possible. The current implementation does this as much as
possible but it still writes to a redis signal queue, the worker
available io and it sets state on the daemon. The concern is that
multiple traps can run at the same time which can cause the system
to get into bad states. The current implementation takes extra
steps to avoid this that can be simplified once the signal handling
is changed.