Update `Process` signal handling
This updates the `Process` signal handling to follow best practices
and do as little as possible when trapping signals. It now simply
writes the signal to an IO pipe. The idea is that you want to avoid
the logic for trapping a signal raising any errors. The previous
implementation didn't do a good job of this because it would change
the daemon's state, write to an IO pipe and write to redis (by
calling `stop` or `halt` on the daemon). This left a greater chance
for it to fail while handling the signal which might cause the
system to get into a bad state. This also makes it better handle
multiple signals being sent to the process and avoid getting in a
bad state from trying to handle them all simultaneously.

This also does some minor improvements to the `Process`. It now
logs when it receives and handles signals. This is to make it more
clear what the process is doing. Its also more robust in that it
won't error if we start using signals that are not supported on
some operating systems.

Finally, this removes the datetime from the bench and system test
logs. The default datetime is very long and makes the logs
annoying to read.