Don't use `gets` and `puts` in `IOPipe`
This changes `IOPipe` from using `gets` and `puts` to instead use
`read_nonblock` and `write_nonblock`. `gets` and `puts` cause
problems because they block. This causes problems and randomly
hangs threads in Ruby 1.8.7. This changes the `IOPipe` to use
`read_nonblock` and `write_nonblock` but also forces writing only
1 byte values.

The `IOPipe` was originally changed to use `gets` and `puts` to
allow the `Process` in a future PR to use it for handling signals.
The signals were going to be written directly to the IO pipe but,
with this change, only a single letter will be written for the
signals.