Change process name, don't allow completely changing it
This changes the process name logic. It will now always prefix
the process name with `qs: `. Furthermore, the env var has been
changed to `QS_PROCESS_LABEL` and it no longer allows completely
changing the process name which is why it was changed. This still
allows the flexibility of changing the process name as needed but
keeps all the processes with the qs prefix.