Rescue redis errors while dequeueing
This updates the logic for dequeueing to rescue any runtime errors
and allow the work loop to continue running. The benefit of this is
if redis goes down or becomes unavailable, it doesn't shut down the
daemon. To keep the process from thrashing, the work loop will
sleep for a second when it errors. Otherwise, if redis is down it
will loop very quickly and error again.

This also improves the daemon logging to be more consistent with
logging elsewhere in Qs and to be more extensive. It will now log
when a worker error occurs.