Shutdown the worker pool if an error occurs in daemon work loop
This changes daemons to shutdown their worker pool when they error.
This is needed to safely shutdown workers that are running and to
avoid losing any work. Previously, the daemon would exit without
shutting down the worker pool which would cause them to be abruptly
exited without any errors being reported.

This also improves the shutdown logging to report that its shutting
down and how long it will allow things to run before it exits. It
also reports how many jobs its requeing as well. This is to be
more verbose and clear about what the daemon is doing.