Restart daemon if its thread shuts down unexpectantly
This updates the `Process` to restart its `Daemon` if the daemon
stops unexpectantly. There is a small chance this could happen
since all redis errors are rescued already but its possible for
the daemon thread to crash and the process will keep running. This
changes it so the process periodically checks in on the daemon as
its listening for signals, if the daemon is no longer running, then
the process restarts it. This avoids the strange case where the
process is running but its not running any jobs because the daemon
has stopped but the `Process` is still waiting for signals.