Simplify error handler unit tests
This simplifies the error handler tests. The tests were overly
complex in their setup and didn't make use of the factories for
generating random sets of data. In particular, there was a lot
of extra handling checking that procs were called and what they
were called with. This encapsulates the logic into an error proc
spy that makes the setup simpler and allows testing that the procs
are called and what they are called with.