Add `Event`
This adds an `Event` object that is built from a job. This will be
used to support events in Qs, both when publishing events and when
processing them. An event composes a job. This is because events are
just special types of jobs and need to make use of a jobs logic.

When publishing an event, an `Event` instance will be built. Since
an `Event` composes a job, publishing the event is as simple as
enqueueing the events job on the dispatchers queue.

Similarly, when processing a job that represents an event, all
the existing logic can be reused because its a job. When an
event handler is built to process the event job, it will use the
job to build an `Event` instance, allowing access to its special
behavior.

This also adds an `Event::JobName` module that will be used to
build event job names when publishing events (via `Event.build`)
and when configuring events on a queue. These need to match up to
ensure events are correctly routed when being processed.

All of this provides the base to implement an events system in Qs
without reworking the existing system and how it handles jobs.