Add `EventHandler` mixin
This adds an `EventHandler` mixin for defining handlers for
processing an event. This is part of setting up an events system
in Qs.

The event handler extends the job handler functionality and uses
the job from its runner to build an `Event`. The params method
on an event handler returns its event params not the job params.
The event handler also provides access to the event itself similar
to how the job handler provides access to the job.