Add `dispatcher_queue` to Qs, setup for publishing events
This adds a `dispatcher_queue` to Qs which will be used when
publishing events. This is part of adding an events system and
more specifically is setup for adding a `publish` method to the
`Qs` module.

The events system works by adding a job to a dispatcher for an
event. The dispatcher then takes the event and adds event jobs to
each queue that is subscribed. To get the initial job to the
dispatcher, Qs needs to know the dispatcher queue to add the job
to and the name of the job for the dispatcher to run. This adds
options for configuring both and exposes them on Qs main module.
This will allow the `Client` to use them to add the event jobs
to the dispatchers queue.