Add payload type to jobs
This updates jobs to know a payload type. This is setup for
supporting events in the system. Specifically, this ensures that
there won't be any collisions when routing jobs and events. With
the previous setup, it would be possible to craft a job/event
combo such that they resulted having the same job name. This would
cause them to not be routed correctly.

This fixes the problem by providing a payload type. Jobs will set
this to "job" and events will set it to "event". When routing the
payload type will be prefixed to the name. This ensures that no
matter what, jobs/events will be routed separately and not collide.

To support this, the `Queue` must also build the prefixed job name
for its routes when they are configured.